requirements.txt
Flask
Flask-Cors
python-dotenv
PyMuPDF
gunicorn
requests
mysql-connector-python
